+++ src/OrchardCore.Modules/OrchardCore.OpenId/Manifest.cs
@@ -1,5 +1,6 @@
 using OrchardCore.Modules.Manifest;
 using OrchardCore.OpenId;
+using OrchardCore.Users;
 
 [assembly: Module(
     Name = "OpenID Connect",
@@ -24,6 +25,7 @@ using OrchardCore.OpenId;
     Dependencies =
     [
         OpenIdConstants.Features.Core,
+        UserConstants.Features.ExternalAuthentication,
     ]
 )]
