+++ src/OrchardCore.Modules/OrchardCore.Users/Extensions/ControllerExtensions.cs
@@ -60,7 +60,7 @@ internal static class ControllerExtensions
 
         var settings = await controller.ControllerContext.HttpContext.RequestServices.GetRequiredService<ISiteService>().GetSettingsAsync<RegistrationSettings>();
 
-        if (settings.UsersCanRegister != UserRegistrationType.NoRegistration)
+        if (!settings.AllowSiteRegistration)
         {
             var registrationEvents = controller.ControllerContext.HttpContext.RequestServices.GetServices<IRegistrationFormEvents>();
