+++ src/OrchardCore.Modules/OrchardCore.Users/Manifest.cs
@@ -20,6 +20,17 @@ using OrchardCore.Users;
 )]
 
 [assembly: Feature(
+    Id = UserConstants.Features.ExternalAuthentication,
+    Name = "External User Authentication",
+    Description = "Provides a way to allow authentication using identity provider.",
+    Dependencies =
+    [
+        UserConstants.Features.Users,
+    ],
+    Category = "Security"
+)]
+
+[assembly: Feature(
     Id = "OrchardCore.Users.ChangeEmail",
     Name = "Users Change Email",
     Description = "The Change email feature allows users to change their email address.",
