+++ src/OrchardCore.Modules/OrchardCore.Users/Models/UserRegistrationType.cs
@@ -1,5 +1,6 @@
 namespace OrchardCore.Users.Models;
 
+[Obsolete("This type is no longer used and will be removed in the next major release.")]
 public enum UserRegistrationType
 {
     NoRegistration = 0,
