+++ src/OrchardCore/OrchardCore.Users.Core/Models/ExternalUserLoginSettings.cs
@@ -1,0 +1,6 @@
+﻿namespace OrchardCore.Users.Models;
+
+public class ExternalUserLoginSettings
+{
+    public bool UseExternalProviderIfOnlyOneDefined { get; set; }
+}
