+++ src/OrchardCore/OrchardCore.Users.Core/Models/ExternalUserRoleLoginSettings.cs
@@ -1,0 +1,8 @@
+﻿namespace OrchardCore.Users.Models;
+
+public class ExternalUserRoleLoginSettings
+{
+    public bool UseScriptToSyncRoles { get; set; }
+
+    public string SyncRolesScript { get; set; }
+}
