+++ src/OrchardCore/OrchardCore.Users.Core/Models/LoginSettings.cs
@@ -6,12 +6,15 @@ public class LoginSettings
 {
     public bool UseSiteTheme { get; set; }
 
+    [Obsolete("This property is no longer used and will be removed in the next major release. Instead use ExternalUserLoginSettings.NoPassword")]
     public bool UseExternalProviderIfOnlyOneDefined { get; set; }
 
     public bool DisableLocalLogin { get; set; }
 
+    [Obsolete("This property is no longer used and will be removed in the next major release. Instead use ExternalUserRoleLoginSettings.SyncRolesScript")]
     public bool UseScriptToSyncRoles { get; set; }
 
+    [Obsolete("This property is no longer used and will be removed in the next major release. Instead use ExternalUserRoleLoginSettings.SyncRolesScript")]
     public string SyncRolesScript { get; set; }
 
     public bool AllowChangingUsername { get; set; }
