+++ src/docs/releases/2.1.0.md
@@ -1,3 +1,24 @@
 # Orchard Core 2.1.0
 
 Release date: Not yet released
+
+## Change Logs
+
+### External User Authentication Feature
+
+A new feature, **External User Authentication**, has been added. This feature was separated from the existing **Users** feature to provide better dependency management and offer an option to disable external authentication by default.
+
+After upgrading the site, a fast-forward option allows you to enable this feature automatically when needed. As a result, the following settings have been relocated to new classes:
+
+- The properties `UseScriptToSyncRoles` and `SyncRolesScript` have been moved from `LoginSettings` to a new class, `ExternalUserRoleLoginSettings`.
+- The property `UseExternalProviderIfOnlyOneDefined` has been moved from `LoginSettings` to `ExternalUserLoginSettings`.
+
+Additionally, the following properties have been moved from `RegistrationSettings` to `ExternalAuthenticationSettings`:
+
+- `NoPasswordForExternalUsers` is now `ExternalAuthenticationSettings.NoPassword`.
+- `NoUsernameForExternalUsers` is now `ExternalAuthenticationSettings.NoUsername`.
+- `NoEmailForExternalUsers` is now `ExternalAuthenticationSettings.NoEmail`.
+- `UseScriptToGenerateUsername` is now `ExternalAuthenticationSettings.UseScriptToGenerateUsername`.
+- `GenerateUsernameScript` is now `ExternalAuthenticationSettings.GenerateUsernameScript`.
+
+The `UsersCanRegister` property in the `RegistrationSettings` class has been marked obsolete, and a new property, `AllowSiteRegistration`, has been introduced. To enable site registration, set `AllowSiteRegistration` to `true`. For external registration, ensure the **External User Authentication** feature is enabled.
